pyproject.toml
[tool.poetry]
name = "prepare-assignment"
version = "0.2.0"
description = "Prepare assignment"
authors = ["Bonajo <m.bonajo@fontys.nl>"]
license = "MIT"
readme = "README.md"
packages = [{include = "prepare_assignment"}]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.9"
ruamel-yaml = "^0.18.5"
jsonschema = "^4.21.1"
importlib-resources = "^6.1.1"
gitpython = "^3.1.41"
prepare-toolbox = "0.4.0"
virtualenv = "^20.25.0"
typer = {extras = ["all"], version = "^0.9.0"}
treelib = "^1.7.0"
multipledispatch = "^1.0.0"
dacite = "^1.9.2"

[tool.poetry.group.test.dependencies]
mypy = "^1.1.1"
types-jsonschema = "^4.17.0.6"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.11.1"

[tool.poetry.scripts]
prepare = 'prepare_assignment.cli.main:app'

[tool.mypy]
warn_return_any = true

[[tool.mypy.overrides]]
module = "treelib,ruamel,virtualenv"
ignore_missing_imports = true